use uid instead of tag index for html id of tag
If you add a tag to an existing point and then delete it, that
tag you added should vanish, not some other tag in the list!  To
ensure this, we need unique IDs for each tag.  We can't rely on
the tags' indices because as you remove tags, the indicies of the
other tags changes.